chore(today): retire V1/V2/V3 input bars and finish Capture v2 cleanup
PR #139 landed the first half of Capture v2 (TodayView empty state
+ removed two API-key banners). This commit finishes the job by
collapsing the four-version input bar zoo and applying Capture v2's
remaining design notes to the canonical InputBarV4.

The variant switch was a feature-flag carcass: each new design just
added another file (InputBarView → V2 → V3 → V4) and bumped the
default. Nobody was switching back, but every change had to be
mirrored across four parallel implementations.

Deleted (~1700 LOC of dead code):
- InputBarView.swift  (V1, legacy)
- InputBarV2.swift    (Fromm-style, US-007)
- InputBarV3.swift    (voice-first, issue #76)
- TodayEmptyStateView.swift (V1's prompt-bubble empty state)
- pbxproj references for the above
- TodayView's `inputBarVariant` / `useInputBarV2` AppStorage + the
  if/else if/else branching that drove them
- SettingsView's variant Picker and `useInputBarV2` toggle

Applied to InputBarV4 (now the canonical input bar):
- Removed the bottom Notes / Attach / Camera shelf. Capture v2
  design note 03: "icon labels are a tell that the icons are not
  legible. Either the icon works on its own, or it should be
  redesigned. We commit." Attachments live behind the `+` button
  inside the composing capsule.
- Split the collapsed capsule into two unambiguous siblings:
    [ "Hold to talk · tap to write" pill ] [ visible mic-hero ]
  Previously the press-to-talk button was an invisible overlay on
  the right quarter of the pill, so users had no visual cue for
  where to press.
- Added `isRecordingTooShort` boundary (sub-1s elapsed AND silent
  waveform). Covers both press-to-talk send and slide-left
  transcribe paths. Silently dropping a sub-second silent recording
  felt rude; we now warn-haptic, cancel, and surface a "再说久一点
  · 至少 1 秒" toast. Brief but audible utterances ("嗯", "对", "好")
  still go through.

Verified by xcodebuild + iPhone 17 simulator visual check —
8 files changed, 107 insertions(+), 1862 deletions(-).
